requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
psycopg2-binary>=2.9
dj-database-url>=2.2
whitenoise>=6.6
django-cors-headers>=4.4
djangorestframework>=3.16
PyJWT>=2.10.1
datetime>=5.5
sib-api-v3-sdk
razorpay
firebase-admin



